webapp/requirements.txt
uvicorn[standard]==0.30.1
fastapi==0.111.0
pyserial==3.5
Jinja2==3.1.4